backend/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
bidict==0.22.0
billiard==3.6.4.0
cached-property==1.5.2
ccxt==1.82.43
celery==5.2.6
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Deprecated==1.2.13
dnspython==2.2.1
eventlet==0.33.1
Flask==2.0.2
Flask-SocketIO==5.2.0
frozenlist==1.3.1
gevent==21.12.0
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
prompt-toolkit==3.0.31
pycares==4.2.2
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
redis==4.1.2
requests==2.27.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0